liveness: shifted around architecture of cfg and related analysis passes
