+++ parsley-core/parsley-core.cabal
@@ -71,8 +71,7 @@ library
 
                        Parsley.Internal.Frontend.Analysis,
                        Parsley.Internal.Frontend.Analysis.Cut,
-                       Parsley.Internal.Frontend.Analysis.Liveness,
-
+ 
                        Parsley.Internal.Frontend.Analysis.Dependencies,
                        Parsley.Internal.Frontend.Analysis.Flags,
                        Parsley.Internal.Frontend.Analysis.Inliner,
