+++ parsley-core/src/ghc/Parsley/Internal/Common/Indexed.hs
@@ -96,3 +96,8 @@ newtype Const1 a k = Const1 {getConst1 :: a}
 
 data Unit4 i j k l = Unit4
 newtype Const4 a i j k l = Const4 {getConst4 :: a}
+
+data Tag t f (k :: Type -> Type) a = Tag {tag :: t, tagged :: f k a}
+
+instance IFunctor f => IFunctor (Tag t f) where
+  imap f (Tag t k) = Tag t (imap f k)
