+++ parsley-core/src/ghc/Parsley/Internal/Frontend/Analysis.hs
