+++ parsley-core/src/ghc/Parsley/Internal/Frontend/Analysis/ReachingDefs.hs
@@ -1,0 +1,7 @@
+module Parsley.Internal.Frontend.Analysis.ReachingDefs (reachingDefAnalysis) where 
+
+import Parsley.Internal.Frontend.Analysis.CFG (CFG)
+
+-- TODO: write this analysis pass
+reachingDefAnalysis :: CFG -> ()
+reachingDefAnalysis = undefined
